// apps/web/src/app/features/inventory/modules/drug-components/types/drug-components.types.ts
export interface DrugComponent {
  id: number;
  generic_id: number;
  component_name: string;
  strength?: string | null;
  strength_value?: number | null;
  strength_unit?: string | null;
  is_active?: boolean | null;
  created_at?: string | null;
  updated_at?: string | null;
}